Refactor Calendar component into smaller, maintainable components
- Split 975-line Calendar component into focused, single-responsibility components
- Create feature-based folder structure with types, utils, hooks, and components
- Extract reusable logic into custom hooks (useCalendar, useWorkouts)
- Implement CSS Modules for all styling (no Tailwind)
- Create separate components for:
  - CalendarNavigation: Month navigation and view mode controls
  - CalendarGrid: Calendar grid with day cells
  - CalendarDay: Individual day cell component
  - CalendarStats: Stats cards section
  - UpcomingWorkouts: Sidebar with upcoming workouts
  - QuickActions: Quick actions sidebar
  - WorkoutEventCard: Individual workout event display
  - WorkoutTypeLegend: Legend for workout types
- Extract utilities for date calculations and workout management
- Improve maintainability, testability, and reusability
- Follow React best practices and code splitting principles
