+++ src/features/calendar/components/CalendarDay/CalendarDay.module.css
@@ -1,0 +1,102 @@
+.day {
+  position: relative;
+  min-height: 5rem;
+  padding: 0.5rem;
+  border: 1px solid #f3f4f6;
+  cursor: pointer;
+  transition: all 0.2s;
+}
+
+.day.currentMonth {
+  background-color: white;
+}
+
+.day.currentMonth:hover {
+  background-color: #f9fafb;
+}
+
+.day.otherMonth {
+  background-color: #f9fafb;
+  color: #9ca3af;
+}
+
+.day.today {
+  background-color: #eef2ff;
+  border-color: #c7d2fe;
+}
+
+.day.selected {
+  ring: 2px;
+  ring-color: #6366f1;
+}
+
+.dateNumber {
+  font-size: 0.875rem;
+  font-weight: 500;
+}
+
+.currentMonthText {
+  color: #111827;
+}
+
+.otherMonthText {
+  color: #9ca3af;
+}
+
+.todayText {
+  color: #4f46e5;
+}
+
+.workoutIndicators {
+  position: absolute;
+  bottom: 0.25rem;
+  left: 0.25rem;
+  right: 0.25rem;
+}
+
+.singleWorkout {
+  display: flex;
+  gap: 0.25rem;
+}
+
+.indicator {
+  width: 0.5rem;
+  height: 0.5rem;
+  border-radius: 50%;
+}
+
+.smallIndicator {
+  width: 0.375rem;
+  height: 0.375rem;
+  border-radius: 50%;
+}
+
+.multipleWorkouts {
+  display: flex;
+  flex-direction: column;
+  gap: 0.125rem;
+}
+
+.indicatorRow {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.indicators {
+  display: flex;
+  gap: 0.125rem;
+}
+
+.moreIndicator {
+  font-size: 0.625rem;
+  color: #6b7280;
+}
+
+.sessionCount {
+  font-size: 0.625rem;
+  color: #6b7280;
+  background-color: rgba(255, 255, 255, 0.8);
+  padding: 0.125rem 0.25rem;
+  border-radius: 0.25rem;
+}
