+++ src/features/calendar/components/CalendarDay/index.ts
@@ -1,0 +1,1 @@
+export { CalendarDay } from "./CalendarDay";
