+++ src/features/calendar/components/CalendarGrid/index.ts
@@ -1,0 +1,1 @@
+export { CalendarGrid } from "./CalendarGrid";
