+++ src/features/calendar/components/CalendarNavigation/index.ts
@@ -1,0 +1,1 @@
+export { CalendarNavigation } from "./CalendarNavigation";
