+++ src/features/calendar/components/CalendarStats/index.ts
@@ -1,0 +1,1 @@
+export { CalendarStats } from "./CalendarStats";
