+++ src/features/calendar/components/QuickActions/QuickActions.module.css
@@ -1,0 +1,26 @@
+.container {
+  padding: 1.5rem;
+}
+
+.title {
+  font-size: 1.125rem;
+  font-weight: 600;
+  color: #1f2937;
+  margin-bottom: 1rem;
+}
+
+.actions {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.actionButton {
+  justify-content: flex-start;
+}
+
+.actionIcon {
+  width: 1rem;
+  height: 1rem;
+  margin-right: 0.75rem;
+}
