+++ src/features/calendar/components/UpcomingWorkouts/UpcomingWorkouts.module.css
@@ -1,0 +1,79 @@
+.container {
+  padding: 1.5rem;
+}
+
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 1.5rem;
+}
+
+.title {
+  font-size: 1.125rem;
+  font-weight: 600;
+  color: #1f2937;
+}
+
+.viewAllButton {
+  color: #4f46e5;
+}
+
+.viewAllButton:hover {
+  color: #4338ca;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.dayGroup {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.dayHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.dayTitle {
+  font-weight: 600;
+  color: #1f2937;
+}
+
+.sessionBadge {
+  font-size: 0.75rem;
+}
+
+.workoutsList {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.emptyState {
+  text-align: center;
+  padding: 2rem 0;
+  color: #6b7280;
+}
+
+.emptyIcon {
+  width: 3rem;
+  height: 3rem;
+  margin: 0 auto 0.75rem;
+  color: #d1d5db;
+}
+
+.emptyTitle {
+  font-weight: 500;
+  margin-bottom: 0.25rem;
+}
+
+.emptyDescription {
+  font-size: 0.875rem;
+}
