+++ src/features/calendar/components/UpcomingWorkouts/index.ts
@@ -1,0 +1,1 @@
+export { UpcomingWorkouts } from "./UpcomingWorkouts";
