+++ src/features/calendar/components/WorkoutEventCard/index.ts
@@ -1,0 +1,1 @@
+export { WorkoutEventCard } from "./WorkoutEventCard";
