+++ src/features/calendar/components/WorkoutTypeLegend/index.ts
@@ -1,0 +1,1 @@
+export { WorkoutTypeLegend } from "./WorkoutTypeLegend";
