+++ src/features/calendar/config/workoutTypes.ts
@@ -1,0 +1,51 @@
+import { WorkoutTypeConfig } from "../types";
+
+export const workoutTypes: WorkoutTypeConfig[] = [
+  {
+    name: "Strength",
+    color: "bg-indigo-500",
+    textColor: "text-indigo-700",
+    bgColor: "bg-indigo-50",
+  },
+  {
+    name: "Cardio",
+    color: "bg-green-500",
+    textColor: "text-green-700",
+    bgColor: "bg-green-50",
+  },
+  {
+    name: "HIIT",
+    color: "bg-orange-500",
+    textColor: "text-orange-700",
+    bgColor: "bg-orange-50",
+  },
+  {
+    name: "Flexibility",
+    color: "bg-purple-500",
+    textColor: "text-purple-700",
+    bgColor: "bg-purple-50",
+  },
+  {
+    name: "Recovery",
+    color: "bg-blue-500",
+    textColor: "text-blue-700",
+    bgColor: "bg-blue-50",
+  },
+];
+
+export const months = [
+  "January",
+  "February",
+  "March",
+  "April",
+  "May",
+  "June",
+  "July",
+  "August",
+  "September",
+  "October",
+  "November",
+  "December",
+];
+
+export const dayHeaders = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
