+++ src/features/calendar/data/mockWorkouts.ts
@@ -1,0 +1,178 @@
+import { WorkoutEvent } from "../types";
+
+export const mockWorkouts: WorkoutEvent[] = [
+  {
+    id: "1",
+    name: "Upper Body Strength",
+    duration: "45 minutes",
+    difficulty: "Intermediate",
+    type: "Strength",
+    status: "completed",
+    date: new Date(2025, 6, 20), // July 20, 2025
+    time: "5:30 PM",
+    exercises: 8,
+    estimatedCalories: 320,
+    actualCalories: 298,
+  },
+  {
+    id: "2",
+    name: "Morning HIIT",
+    duration: "30 minutes",
+    difficulty: "Advanced",
+    type: "HIIT",
+    status: "scheduled",
+    date: new Date(2025, 6, 22), // July 22, 2025 (Today)
+    time: "7:00 AM",
+    exercises: 6,
+    estimatedCalories: 280,
+  },
+  {
+    id: "2b",
+    name: "Evening Yoga",
+    duration: "25 minutes",
+    difficulty: "Beginner",
+    type: "Flexibility",
+    status: "scheduled",
+    date: new Date(2025, 6, 22), // July 22, 2025 (Today - Evening)
+    time: "7:30 PM",
+    exercises: 4,
+    estimatedCalories: 120,
+  },
+  {
+    id: "3",
+    name: "Lower Body Focus",
+    duration: "50 minutes",
+    difficulty: "Intermediate",
+    type: "Strength",
+    status: "scheduled",
+    date: new Date(2025, 6, 23), // July 23, 2025 (Tomorrow)
+    time: "6:00 PM",
+    exercises: 10,
+    estimatedCalories: 350,
+  },
+  {
+    id: "4a",
+    name: "Morning Walk",
+    duration: "30 minutes",
+    difficulty: "Beginner",
+    type: "Cardio",
+    status: "scheduled",
+    date: new Date(2025, 6, 24), // July 24, 2025
+    time: "7:00 AM",
+    exercises: 2,
+    estimatedCalories: 150,
+  },
+  {
+    id: "4b",
+    name: "Strength Training",
+    duration: "45 minutes",
+    difficulty: "Intermediate",
+    type: "Strength",
+    status: "scheduled",
+    date: new Date(2025, 6, 24), // July 24, 2025 (Same day - Evening)
+    time: "6:30 PM",
+    exercises: 7,
+    estimatedCalories: 300,
+  },
+  {
+    id: "5",
+    name: "Full Body Circuit",
+    duration: "40 minutes",
+    difficulty: "Advanced",
+    type: "HIIT",
+    status: "completed",
+    date: new Date(2025, 6, 18), // July 18, 2025
+    time: "6:30 PM",
+    exercises: 12,
+    estimatedCalories: 380,
+    actualCalories: 365,
+  },
+  {
+    id: "6a",
+    name: "Morning Cardio",
+    duration: "25 minutes",
+    difficulty: "Beginner",
+    type: "Cardio",
+    status: "completed",
+    date: new Date(2025, 6, 19), // July 19, 2025
+    time: "6:45 AM",
+    exercises: 3,
+    estimatedCalories: 180,
+    actualCalories: 165,
+  },
+  {
+    id: "6b",
+    name: "Core & Flexibility",
+    duration: "35 minutes",
+    difficulty: "Beginner",
+    type: "Flexibility",
+    status: "completed",
+    date: new Date(2025, 6, 19), // July 19, 2025 (Same day - Evening)
+    time: "8:00 PM",
+    exercises: 6,
+    estimatedCalories: 150,
+    actualCalories: 142,
+  },
+  {
+    id: "7a",
+    name: "Morning Run",
+    duration: "35 minutes",
+    difficulty: "Intermediate",
+    type: "Cardio",
+    status: "completed",
+    date: new Date(2025, 6, 21), // July 21, 2025
+    time: "6:30 AM",
+    exercises: 3,
+    estimatedCalories: 250,
+    actualCalories: 265,
+  },
+  {
+    id: "7b",
+    name: "Push Day",
+    duration: "55 minutes",
+    difficulty: "Advanced",
+    type: "Strength",
+    status: "completed",
+    date: new Date(2025, 6, 21), // July 21, 2025 (Same day - Evening)
+    time: "6:00 PM",
+    exercises: 9,
+    estimatedCalories: 380,
+    actualCalories: 395,
+  },
+  {
+    id: "8",
+    name: "Morning Cardio",
+    duration: "25 minutes",
+    difficulty: "Intermediate",
+    type: "Cardio",
+    status: "scheduled",
+    date: new Date(2025, 6, 25), // July 25, 2025
+    time: "6:30 AM",
+    exercises: 5,
+    estimatedCalories: 220,
+  },
+  {
+    id: "9a",
+    name: "Lunch Break Stretch",
+    duration: "20 minutes",
+    difficulty: "Beginner",
+    type: "Flexibility",
+    status: "scheduled",
+    date: new Date(2025, 6, 26), // July 26, 2025
+    time: "12:30 PM",
+    exercises: 4,
+    estimatedCalories: 80,
+  },
+  {
+    id: "9b",
+    name: "Pull Day",
+    duration: "50 minutes",
+    difficulty: "Intermediate",
+    type: "Strength",
+    status: "scheduled",
+    date: new Date(2025, 6, 26), // July 26, 2025 (Same day - Evening)
+    time: "7:00 PM",
+    exercises: 8,
+    estimatedCalories: 340,
+  },
+];
