+++ src/features/calendar/index.ts
@@ -1,0 +1,8 @@
+export { default as Calendar } from "./Calendar";
+export * from "./types";
+export * from "./hooks/useCalendar";
+export * from "./hooks/useWorkouts";
+export * from "./utils/calendarUtils";
+export * from "./utils/workoutUtils";
+export * from "./config/workoutTypes";
+export * from "./data/mockWorkouts";
